requirements.txt
aiosmtpd==1.4.4.post2
atpublic==4.0
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.6
